refactor: clean code and reused functions
