+++ apps/bot/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './getComponent';
+export * from './getFilesRecursively';
 export * from './loadCommands';
 export * from './loadEvents';
