+++ apps/bot/src/utils/loadCommands.ts
@@ -1,33 +1,13 @@
-import fs from 'fs';
 import path from 'path';
 import { Command } from '../common';
+import { getFilesRecursively } from './getFilesRecursively';
 import { Client, Collection } from 'discord.js';
-
-// Function to recursively get all .ts files in a directory
-const getCommandFilesRecursively = (dir: string): Array<string> => {
-  let results: Array<string> = [];
-  const list = fs.readdirSync(dir);
-
-  list.forEach((file) => {
-    const filePath = path.join(dir, file);
-    const stat = fs.statSync(filePath);
-
-    if (stat && stat.isDirectory()) {
-      // Recursively search subdirectories
-      results = results.concat(getCommandFilesRecursively(filePath));
-    } else if (file.endsWith('.ts')) {
-      results.push(filePath);
-    }
-  });
-
-  return results;
-};
 
 export const loadCommands = async (client: Client) => {
   client.commands = new Collection();
 
   const commandsPath = path.join(__dirname, '..', 'commands');
-  const commandFiles = getCommandFilesRecursively(commandsPath);
+  const commandFiles = getFilesRecursively(commandsPath);
 
   for (const file of commandFiles) {
     const command = (await import(file)).default as Command;
